Extract API key crypto helpers
